refactor: set the right vip number and the description for VIP-163
